Trim movie overview and confirmation message for favorite removal

+++ src/app/favorites/components/favorite-preview/favorite-preview.component.ts
@@ -1,6 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { Store } from '@ngrx/store';
+import { Observable } from 'rxjs';
 import { Movie } from 'src/app/home/interfaces/movies.interface';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 import { FavoriteActions } from '../../store/favorites-actions';
 
 @Component({
@@ -8,14 +11,26 @@ import { FavoriteActions } from '../../store/favorites-actions';
   templateUrl: './favorite-preview.component.html',
   styleUrls: ['./favorite-preview.component.scss'],
 })
-export class FavoritePreviewComponent implements OnInit {
+export class FavoritePreviewComponent {
   @Input() movie!: Movie;
+  confirmDialog! : MatDialogRef<ConfirmationDialogComponent>
 
-  constructor(private store: Store) {}
+  constructor(private store: Store, private dialog: MatDialog) {}
 
-  ngOnInit(): void {}
+  openConfirmationDialog(): Observable<boolean> {
+    this.confirmDialog = this.dialog.open(ConfirmationDialogComponent, {
+      disableClose: true,
+    });
+    this.confirmDialog.componentInstance.dialogMessage =
+      'Are you sure you want to remove this movie from your favorites?';
+    return this.confirmDialog.afterClosed();
+  }
 
   removeFromFavorites() : void {
-    this.store.dispatch(FavoriteActions.deleteFavorite({favoriteMovieId: this.movie.id}))
+    this.openConfirmationDialog().subscribe(result => {
+      if(result){
+        this.store.dispatch(FavoriteActions.deleteFavorite({favoriteMovieId: this.movie.id}))
+      }
+    })
   }
 }
